# Parallelize optimized and unoptimized bytecode reports in CI