~Depends on #14353.~ Merged.
Closes #13056.

This is the second step in parallelizing the bytecode report generation, after #14330. Now that I can generate optimized and unoptimized report separately via presets (#14350), I'm doing just that, in parallel.

This time using CircleCI's job matrix rather than `circleci tests split` utility. The downside is that it doubles the number of jobs. The upside is that it's simpler.