@cameel 2023-06-23
I adjusted the repo settings to make the `b_bytecode_*` jobs not required. Without it, this PR would not be mergeable.

Note that `t_bytecode_compare` is still required and depends on all those jobs, so a PR that fails to generate any of the reports won't pass required checks anyway.