@cameel 2023-06-23
Here's the timing before and after:

### Before
| Job                     | duration |
|-------------------------|----------|
| `b_bytecode_win`        | 11m 24s  |
| `b_bytecode_osx`        | 6m 8s    |
| `b_bytecode_ems`        | 7m 21s   |
| `b_bytecode_ubu_static` | 2m 57s   |
| `b_bytecode_ubu`        | 5m 51s   |

### After
| Job                                        | duration |
|--------------------------------------------|----------|
| `b_bytecode_win-legacy-optimize`           | 6m 45s   |
| `b_bytecode_win-legacy-no-optimize`        | 4m 46s   |
| `b_bytecode_osx-legacy-optimize`           | 4m 13s   |
| `b_bytecode_osx-legacy-no-optimize`        | 3m 8s    |
| `b_bytecode_ems-legacy-optimize`           | 5m 2s    |
| `b_bytecode_ems-legacy-no-optimize`        | 2m 8s    |
| `b_bytecode_ubu_static-legacy-optimize`    | 2m 22s   |
| `b_bytecode_ubu_static-legacy-no-optimize` | 1m 18s   |
| `b_bytecode_ubu-legacy-optimize`           | 3m 52s   |
| `b_bytecode_ubu-legacy-no-optimize`        | 2m 48s   |